Experimenting with output

+++ tests/lbaas_utils.py
@@ -210,15 +210,15 @@ def validate_metering(lb_test_case, requests, total_bytes):
     logging.info("             : %s GB" %convert_to_gb(total_bytes))
     # compare test-detected bytes to metered bytes...
     logging.info("Comparing test-tool received bytes to metered bytes...")
-    metered_bytes = long(metered_bytes)
-    total_bytes = long(total_bytes)
-    byte_diff = long(total_bytes - metered_bytes)
+    metered_bytes = float(metered_bytes)
+    total_bytes = float(total_bytes)
+    byte_diff = float(total_bytes - metered_bytes)
     if byte_diff:
         logging.info("Difference between metered and expected bytes: %s" %byte_diff)
         logging.info("                                             : %s GB" %convert_to_gb(byte_diff))
         per_req_diff = float(byte_diff/requests)
         logging.info("Difference per request: %s" %per_req_diff)
-        percent_diff = float(byte_diff)/float(total_bytes)*100
+        percent_diff = float(byte_diff/total_bytes)*100
         logging.info("Percentage difference: %s" %percent_diff)
     # scan messages for format ...
     # scan messages for type...
